Add clang-tidy to CI
Phase 2 track 4. .clang-tidy selects a narrow, bug-focused check set
rather than the broad modernize-*/readability-* families, which would
flag thousands of lines of 2005-era style with no correctness benefit.
Every exclusion is justified in the file against the actual findings.

One fix: the connect timeout computed tv_usec as 1000 * (msec % 1000)
in int arithmetic before widening to the long that timeval expects.
Not reachable as an overflow given the operand ranges, but the cast
costs nothing and lets the check stay on for cases that would be.

Nothing else survived review. The eight cert-err33-c findings are all
snprintf() into buffers explicitly sized for the message, and the
single branch-clone is intentional; both are excluded with that
reasoning recorded.

CI gains a format job and a tidy job. .clang-tidy sets
WarningsAsErrors, so any new finding fails the build.

+++ src/socket_connect_normal.cpp
@@ -133,7 +133,7 @@ int Socket_Connect_Normal::connecttimeout(int socket, struct sockaddr *addr, soc
 		FD_SET(socket, &wset);
 
 		tv.tv_sec = msec / 1000;
-		tv.tv_usec = 1000 * (msec % 1000);
+		tv.tv_usec = (long) 1000 * (msec % 1000);
 
 		ret = select(socket + 1, NULL, &wset, NULL, &tv);
